Separate configuration preset definitions
